requirements.txt
fastapi
uvicorn
discord.py
python-dotenv
apscheduler
aiofiles